requirements.txt
# Core ML frameworks
tensorflow>=2.6.0
tensorflow-federated
torch>=1.9.0
transformers

# Data processing
pandas>=1.3.0
numpy>=1.19.0
scikit-learn

# RAG components
elasticsearch
faiss-cpu

# Privacy and security
tensorflow-privacy
pysyft

# API and web
flask>=2.0.0
fastapi
uvicorn
requests>=2.25.0

# Configuration and utilities
pyyaml>=5.4.0

# Testing and development
pytest
black
flake8
isort

# Documentation
sphinx
sphinx-rtd-theme

# Additional requirements
pyyaml>=5.4.1
